requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.5
django-appconf==1.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
pilkit==2.0
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
